backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
crewai
httpx
python-dotenv
supabase
